Fixing bug w/ Exec overwriting siteSettings

+++ Exec.ps1
@@ -93,7 +93,10 @@ UI_ADMIN_DOCKER=$($UiFrontDocker)
 
 	Write-Host "Initializing container.."
 
-	docker exec -t $WebContainer cp docker/siteSettings.json siteSettings.json
+	if (!(Test-Path -Path siteSettings.json)) {
+ 		docker exec -t $WebContainer cp docker/siteSettings.json siteSettings.json
+	}
+
 	docker exec -t $WebContainer composer update
 	docker exec -t $WebContainer vendor/bin/stoic-migrate up
